requirements.txt
Django==1.8.6
Markdown==2.6.4
django-filter==0.11.0
django-require==1.0.8
djangorestframework==3.3.1
wsgiref==0.1.2
